Disable SysTick interrupt
The SysTick interrupt was enabled by the jgaurora bootloader and stays enabled when it has jumped to the application. This causes issues because the interrupt handling routine increments specific RAM locations.
When compiling the jgaurora A1 with a non-maple configuration this causes problems because this specific memory location is reused for GPIOA_BASE. Effectively it becomes unable to use GPIOA.
Another side effect is that the bootloader also puts some function pointers in RAM and will periodically call those functions. When this happens after datainit or zerobss (with enough memory to init/zero) it will result in a hard fault exception.